Updated ngwebdriver and change to waitForAngularRequestsToFinish API, moved script timeout to WebDriverGuiceModule, also changed some BasePage methods to protected not public e.g. executeJS.

+++ src/main/java/com/frameworkium/core/ui/pages/PageFactory.java
@@ -13,7 +13,7 @@ public class PageFactory {
     }
 
     public static <T extends BasePage<T>> T newInstance(
-            Class<T> clazz, Integer timeoutInSeconds) {
+            Class<T> clazz, long timeoutInSeconds) {
         return Guice
                 .createInjector(new WebDriverGuiceModule())
                 .getInstance(clazz)
@@ -29,7 +29,7 @@ public class PageFactory {
     }
 
     public static <T extends BasePage<T>> T newInstance(
-            Class<T> clazz, String url, Integer timeoutInSeconds) {
+            Class<T> clazz, String url, long timeoutInSeconds) {
         return Guice
                 .createInjector(new WebDriverGuiceModule())
                 .getInstance(clazz)
